Moved functions out of details into more appropriate locations. Renamed slice to utils to reflect util functions living. Added an option configuration to skip fine tuning.
